fix: scope extension refactor source dispatch
